+++ service/controller/rbac/resource/externalresources/resource.go
@@ -4,7 +4,7 @@
 package externalresources
 
 import (
-	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
+	"github.com/giantswarm/k8sclient/v8/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	rbacv1 "k8s.io/api/rbac/v1"