+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2026-07-21
+
 ### Added
 
 - Create a per-organization ClusterRoleBinding granting the org's `automation` ServiceAccount the `kamaji-datastore-manager` ClusterRole, so cluster apps using Kamaji can manage their `kamaji.clastix.io/datastores` CR. The referenced ClusterRole is provisioned by the global Kamaji app.
@@ -591,7 +593,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Read-only role for customer access into Control Plane.
 
-[Unreleased]: https://github.com/giantswarm/rbac-operator/compare/v0.44.0...HEAD
+[Unreleased]: https://github.com/giantswarm/rbac-operator/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/giantswarm/rbac-operator/compare/v0.44.0...v1.0.0
 [0.44.0]: https://github.com/giantswarm/rbac-operator/compare/v0.43.0...v0.44.0
 [0.43.0]: https://github.com/giantswarm/rbac-operator/compare/v0.42.4...v0.43.0
 [0.42.4]: https://github.com/giantswarm/rbac-operator/compare/v0.42.3...v0.42.4